Pulled all functionality out into functions if possible in charts
